requirements.txt
Flask==3.0.3
APScheduler==3.10.4
SQLAlchemy==2.0.30
requests==2.32.3
python-dotenv==1.0.1
email-validator==2.1.1
pytz==2024.1